split out experiences components into own files and remove testing styles
