+++ src/components/Experiences/index.ts
@@ -1,0 +1,2 @@
+export { default as Details } from './details';
+export { default as Menu } from './menu';
